Revert "Feature/add jwt token and encryption"
