+++ package.json
@@ -17,10 +17,8 @@
     "test": "jest",
     "test:watch": "jest --watch",
     "test:cov": "jest --coverage",
-    "verify": "yarn run lint && yarn run test:cov  --passWithNoTests && yarn run test:e2e && ts-node ./merge-coverage.ts --report coverage/unit/coverage-final.json --report coverage/e2e/coverage-final.json",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
     "test:e2e": "jest --config ./test/jest-e2e.json",
-    "test:e2e:cov": "jest --config ./test/jest-e2e.json --runInBand --detectOpenHandles --coverage",
     "migration:generate": "ts-node node_modules/.bin/typeorm migration:generate --config ./ormconfig.js --name",
     "migration:run": "ts-node node_modules/.bin/typeorm migration:run --config ./ormconfig.js",
     "migration:revert": "ts-node node_modules/.bin/typeorm migration:revert --config ./ormconfig.js",
@@ -29,23 +27,17 @@
   },
   "dependencies": {
     "@nestjs/common": "^7.6.15",
-    "@nestjs/config": "^0.6.3",
     "@nestjs/core": "^7.6.15",
     "@nestjs/graphql": "^7.11.0",
-    "@nestjs/jwt": "^7.2.0",
-    "@nestjs/passport": "^7.1.6",
     "@nestjs/platform-express": "^7.6.15",
     "@nestjs/typeorm": "^7.1.5",
     "apollo-server-express": "^2.25.2",
-    "bcrypt": "^5.0.1",
     "class-validator": "^0.13.1",
-    "dotenv": "^10.0.0",
     "graphql": "^15.5.1",
     "graphql-tools": "^7.0.5",
     "nestjs-typeorm-paginate": "^2.5.3",
-    "passport": "^0.4.1",
-    "passport-jwt": "^4.0.0",
     "pg": "^8.6.0",
+    "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
     "rxjs": "^6.6.6",
     "typeorm": "^0.2.34",
@@ -59,15 +51,12 @@
     "@types/faker": "^5.5.6",
     "@types/jest": "^26.0.22",
     "@types/node": "^14.14.36",
-    "@types/passport-jwt": "^3.0.5",
     "@types/supertest": "^2.0.10",
     "@typescript-eslint/eslint-plugin": "^4.19.0",
     "@typescript-eslint/parser": "^4.19.0",
     "eslint": "^7.22.0",
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-prettier": "^3.3.1",
-    "fs-extra": "^10.0.0",
-    "istanbul-api": "^3.0.0",
     "jest": "^26.6.3",
     "prettier": "^2.2.1",
     "supertest": "^6.1.3",
@@ -84,22 +73,14 @@
       "ts"
     ],
     "rootDir": "src",
-    "testRegex": ".spec.ts$",
+    "testRegex": ".*\\.spec\\.ts$",
     "transform": {
       "^.+\\.(t|j)s$": "ts-jest"
     },
-    "coverageReporters": [
-      "json",
-      "lcov",
-      "text",
-      "html"
+    "collectCoverageFrom": [
+      "**/*.(t|j)s"
     ],
-    "coverageDirectory": "../coverage/unit",
-    "coveragePathIgnorePatterns": [
-      ".module.ts$",
-      ".spec.ts$",
-      "src/main.ts"
-    ],
+    "coverageDirectory": "../coverage",
     "testEnvironment": "node"
   }
 }
