+++ schema.gql
@@ -2,6 +2,15 @@
 # THIS FILE WAS AUTOMATICALLY GENERATED (DO NOT MODIFY)
 # ------------------------------------------------------
 
+type User {
+  id: String!
+  email: String!
+  full_name: String!
+  photo_path: String!
+  password_salt: String!
+  password_hash: String!
+}
+
 type Video {
   id: Float!
   size: String!
@@ -28,41 +37,12 @@ type Movie {
   imdb_id: String!
   videos: [Video!]
   genres: [Genre!]
-  lists: [List!]
 }
 
 """
 A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format.
 """
 scalar DateTime
-
-type List {
-  id: Float!
-  name: String!
-  description: String!
-  public: Boolean!
-  userId: Int!
-  user: User
-  movies: [Movie!]
-}
-
-type User {
-  id: Float!
-  email: String!
-  full_name: String!
-  photo_path: String!
-  password_salt: String!
-  password_hash: String!
-  lists: [List!]
-}
-
-type Token {
-  access_token: String!
-}
-
-type ResponseStatus {
-  status: String!
-}
 
 type Meta {
   itemCount: Float!
@@ -79,21 +59,19 @@ type PaginatedMovies {
 
 type Query {
   getUsers: [User!]!
-  getUser: User!
+  getUser(id: String!): User!
   getMovies(order: String = "DESC", search: String = "", page: Float = 1, limit: Float = 10): PaginatedMovies!
   getMovie(id: Float!): Movie
   getVideos: [Video!]!
   getVideo(id: Float!): Video!
   getGenres: [Genre!]!
   getGenre(id: Float!): Genre!
-  getLists: [List!]!
-  getList(id: Float!): List!
 }
 
 type Mutation {
-  login(input: LoginInput!): Token!
-  createUser(input: CreateUserInput!): User!
-  updateUser(input: UpdateUserInput!): ResponseStatus!
+  createUser(input: UserInput!): User!
+  updateUser(input: UserInput!): User!
+  deleteUser(id: String!): User
   createMovie(input: MovieInput!): Movie!
   updateMovie(input: MovieInput!): Movie!
   deleteMovie(id: String!): Movie
@@ -102,28 +80,15 @@ type Mutation {
   deleteVideo(id: String!): Video
   createGenre(input: GenreInput!): Genre!
   deleteGenre(id: String!): Genre
-  createList(input: CreateListInput!): List!
-  updateList(input: UpdateListInput!): ResponseStatus!
-  deleteList(id: Float!): ResponseStatus!
-  addMovieToList(input: ListMovieInput!): ResponseStatus!
-  removeMovieToList(input: ListMovieInput!): ResponseStatus!
 }
 
-input LoginInput {
-  email: String!
-  password: String!
-}
-
-input CreateUserInput {
+input UserInput {
+  id: String
   email: String!
   full_name: String!
   photo_path: String!
-  password: String!
-}
-
-input UpdateUserInput {
-  full_name: String!
-  photo_path: String!
+  password_salt: String!
+  password_hash: String!
 }
 
 input MovieInput {
@@ -149,22 +114,4 @@ input VideoInput {
   type: String!
   url: String!
   movieId: Int!
-}
-
-input CreateListInput {
-  name: String!
-  description: String
-  public: Boolean = true
-}
-
-input UpdateListInput {
-  id: Float!
-  name: String!
-  description: String!
-  public: Boolean!
-}
-
-input ListMovieInput {
-  listId: Float!
-  movieId: Float!
 }
