--- src/auth/auth.constants.ts
@@ -1,11 +1,0 @@
-import { ConfigService } from '@nestjs/config';
-
-export const jwtConstants = {
-  inject: [ConfigService],
-  useFactory: (config: ConfigService) => ({
-    secret: config.get<string>('tokenAuth'),
-    signOptions: { expiresIn: '1d' },
-  }),
-};
-
-export const whiteListResolver = ['login', 'createUser'];
