--- src/auth/auth.decorator.ts
@@ -1,9 +1,0 @@
-import { GqlExecutionContext } from '@nestjs/graphql';
-import { createParamDecorator, ExecutionContext } from '@nestjs/common';
-
-export const CurrentUser = createParamDecorator(
-  (data: unknown, context: ExecutionContext) => {
-    const ctx = GqlExecutionContext.create(context);
-    return ctx.getContext().req.user;
-  },
-);
