--- src/lists/dto/args/get-list.arg.ts
@@ -1,9 +1,0 @@
-import { IsNotEmpty } from 'class-validator';
-import { ArgsType, Field } from '@nestjs/graphql';
-
-@ArgsType()
-export class GetListArgs {
-  @Field()
-  @IsNotEmpty()
-  id: number;
-}
