--- src/lists/dto/input/listMovie.input.ts
@@ -1,10 +1,0 @@
-import { Field, InputType } from '@nestjs/graphql';
-
-@InputType()
-export class ListMovieInput {
-  @Field()
-  listId: number;
-
-  @Field()
-  movieId: number;
-}
