--- src/lists/dto/input/updateList.input.ts
@@ -1,16 +1,0 @@
-import { Field, InputType } from '@nestjs/graphql';
-
-@InputType()
-export class UpdateListInput {
-  @Field()
-  id: number;
-
-  @Field()
-  name: string;
-
-  @Field()
-  description: string;
-
-  @Field()
-  public: boolean;
-}
