--- src/lists/lists.repository.ts
@@ -1,6 +1,0 @@
-import { EntityRepository, Repository } from 'typeorm';
-
-import { List } from './lists.entity';
-
-@EntityRepository(List)
-export class ListRepository extends Repository<List> {}
