+++ src/main.ts
@@ -3,6 +3,6 @@ import { AppModule } from './app.module';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
-  await app.listen(AppModule.port);
+  await app.listen(3000);
 }
 bootstrap();
