+++ src/movies/dto/types/movies.paginated.ts
@@ -1,7 +1,7 @@
 import { ObjectType } from '@nestjs/graphql';
 
+import { Paginated } from './paginated';
 import { Movie } from './../../..//movies/movies.entity';
-import { Paginated } from '../../../utils/types/paginated';
 
 @ObjectType()
 export class PaginatedMovies extends Paginated(Movie) {}
