+++ src/movies/movies.module.ts
@@ -7,6 +7,5 @@ import { MovieRepository } from './movies.repository';
 @Module({
   imports: [TypeOrmModule.forFeature([MovieRepository])],
   providers: [MoviesResolver, MoviesService],
-  exports: [MoviesService],
 })
 export class MoviesModule {}
