+++ src/movies/movies.resolver.ts
@@ -1,5 +1,4 @@
 import { Pagination } from 'nestjs-typeorm-paginate';
-import { UseGuards } from '@nestjs/common';
 import { Args, Mutation, Query, Resolver } from '@nestjs/graphql';
 
 import { Movie } from './movies.entity';
@@ -8,25 +7,21 @@ import { MovieInput } from './dto/input/movie.input';
 import { GetMovieArgs } from './dto/args/get-movie.arg';
 import { GetMoviesArgs } from './dto/args/get-movies.arg';
 import { PaginatedMovies } from './dto/types/movies.paginated';
-import { GqlAuthGuard } from './../auth/guards/gpl-auth.guard';
 @Resolver(() => Movie)
 export class MoviesResolver {
   constructor(private readonly _MoviesService: MoviesService) {}
 
   @Query(() => PaginatedMovies)
-  @UseGuards(GqlAuthGuard)
   getMovies(@Args() filters: GetMoviesArgs): Promise<Pagination<Movie>> {
     return this._MoviesService.paginate(filters);
   }
 
   @Query(() => Movie, { nullable: true })
-  @UseGuards(GqlAuthGuard)
   getMovie(@Args() filters: GetMovieArgs): Promise<Movie> {
     return this._MoviesService.getMovie(filters);
   }
 
   @Mutation(() => Movie)
-  @UseGuards(GqlAuthGuard)
   createMovie(
     @Args('input')
     input: MovieInput,
@@ -35,7 +30,6 @@ export class MoviesResolver {
   }
 
   @Mutation(() => Movie)
-  @UseGuards(GqlAuthGuard)
   updateMovie(
     @Args('input')
     input: MovieInput,
@@ -44,7 +38,6 @@ export class MoviesResolver {
   }
 
   @Mutation(() => Movie, { nullable: true })
-  @UseGuards(GqlAuthGuard)
   deleteMovie(
     @Args('id')
     id: string,
