--- src/users/dto/input/login.input.ts
@@ -1,10 +1,0 @@
-import { Field, InputType } from '@nestjs/graphql';
-
-@InputType()
-export class LoginInput {
-  @Field()
-  email: string;
-
-  @Field()
-  password: string;
-}
