--- src/users/dto/input/updateUser.input.ts
@@ -1,10 +1,0 @@
-import { Field, InputType } from '@nestjs/graphql';
-
-@InputType()
-export class UpdateUserInput {
-  @Field()
-  full_name: string;
-
-  @Field()
-  photo_path: string;
-}
