--- src/users/dto/types/token.ts
@@ -1,7 +1,0 @@
-import { Field, ObjectType } from '@nestjs/graphql';
-
-@ObjectType()
-export class Token {
-  @Field()
-  access_token: string;
-}
