--- src/utils/types/response.ts
@@ -1,7 +1,0 @@
-import { Field, ObjectType } from '@nestjs/graphql';
-
-@ObjectType()
-export class ResponseStatus {
-  @Field()
-  status: 'success' | 'fail';
-}
