+++ test/jest-e2e.json
@@ -1,19 +1,9 @@
 {
   "moduleFileExtensions": ["js", "json", "ts"],
-  "rootDir": "..",
-  "testMatch": ["**/*.e2e-spec.ts"],
-  "testPathIgnorePatterns": ["/node_modules/", "/src/"],
+  "rootDir": ".",
+  "testEnvironment": "node",
+  "testRegex": ".e2e-spec.ts$",
   "transform": {
     "^.+\\.(t|j)s$": "ts-jest"
-  },
-  "coverageReporters": ["json", "lcov", "text", "html"],
-  "coverageDirectory": "coverage/e2e",
-  "collectCoverageFrom": ["src/**"],
-  "coveragePathIgnorePatterns": [
-    ".module.ts$",
-    ".spec.ts$",
-    "src/main.ts",
-    "src/database/*"
-  ],
-  "testEnvironment": "node"
+  }
 }
